+++ Assets/Live2D/Cubism/Editor/Importers/CubismModel3JsonImporter.cs
@@ -11,6 +11,7 @@ using Live2D.Cubism.Framework;
 using Live2D.Cubism.Framework.Expression;
 using Live2D.Cubism.Framework.Json;
 using Live2D.Cubism.Framework.MotionFade;
+using Live2D.Cubism.Framework.MouthMovement;
 using Live2D.Cubism.Framework.Pose;
 using Live2D.Cubism.Rendering;
 using System;
@@ -261,12 +262,31 @@ namespace Live2D.Cubism.Editor.Importers
                 // Reset moc reference.
                 CubismModel.ResetMocReference(model, MocAsset);
 
+                // Update moc asset before saving prefab to avoid IndexOutOfRangeException
+                // when Inspector loads the prefab and calls Revive() with stale moc data.
+                if (MocAsset != null)
+                {
+                    EditorUtility.CopySerialized(moc, MocAsset);
+
+                    // Revive by force to make instance using the new Moc.
+                    CubismMoc.ResetUnmanagedMoc(MocAsset);
+
+                    EditorUtility.SetDirty(MocAsset);
+                }
+
                 // Keep layer value.
                 model.gameObject.layer = ModelPrefab.layer;
 
                 // Replace prefab.
 #if UNITY_2018_3_OR_NEWER
                 ModelPrefab = PrefabUtility.SaveAsPrefabAsset(model.gameObject, $"{assetPath}.prefab");
+
+                // Clear stale non-serialized state cached during prefab replacement.
+                var savedModel = ModelPrefab.FindCubismModel();
+                if (savedModel != null)
+                {
+                    CubismModel.ResetNonSerializedFields(savedModel);
+                }
 #else
                 ModelPrefab = PrefabUtility.ReplacePrefab(model.gameObject, ModelPrefab, ReplacePrefabOptions.ConnectToPrefab);
 #endif
@@ -278,20 +298,6 @@ namespace Live2D.Cubism.Editor.Importers
 
             // Clean up.
             Object.DestroyImmediate(model.gameObject, true);
-
-
-            // Update moc asset.
-            if (MocAsset != null)
-            {
-                EditorUtility.CopySerialized(moc, MocAsset);
-
-
-                // Revive by force to make instance using the new Moc.
-                CubismMoc.ResetUnmanagedMoc(MocAsset);
-
-
-                EditorUtility.SetDirty(MocAsset);
-            }
 
             // Save state and assets.
             if (isImporterDirty)
@@ -328,7 +334,8 @@ namespace Live2D.Cubism.Editor.Importers
                 || sourceComponent.GetType() == typeof(CubismFadeController)
                 || sourceComponent.GetType() == typeof(CubismExpressionController)
                 || sourceComponent.GetType() == typeof(CubismPoseController)
-                || sourceComponent.GetType() == typeof(CubismParameterStore))
+                || sourceComponent.GetType() == typeof(CubismParameterStore)
+                || sourceComponent.GetType() == typeof(CubismDisplayInfoCombinedParameterInfo))
                 {
                     continue;
                 }
@@ -375,6 +382,12 @@ namespace Live2D.Cubism.Editor.Importers
                         continue;
                     }
 
+                    // Skip import-managed components that should not be inherited from the old prefab.
+                    if (sourceComponent.GetType() == typeof(CubismEyeBlinkParameter)
+                    ||  sourceComponent.GetType() == typeof(CubismMouthParameter))
+                    {
+                        continue;
+                    }
 
                     // Copy component.
                     var destinationComponent = destinationT.GetOrAddComponent(sourceComponent.GetType());
@@ -383,12 +396,14 @@ namespace Live2D.Cubism.Editor.Importers
                         var name = cdiParameterName.Name;
                         EditorUtility.CopySerialized(sourceComponent, destinationComponent);
                         cdiParameterName.Name = name;
+                        EditorUtility.SetDirty(cdiParameterName);
                     }
                     else if (destinationComponent is CubismDisplayInfoPartName cdiPartName && !string.IsNullOrEmpty(cdiPartName.Name))
                     {
                         var name = cdiPartName.Name;
                         EditorUtility.CopySerialized(sourceComponent, destinationComponent);
                         cdiPartName.Name = name;
+                        EditorUtility.SetDirty(cdiPartName);
                     }
                     else
                     {